Move constants and restructure directory - constants moved to constants.py file - create an overall "bot" directory - renamed lovelace.py to bot.py
