+++ bot/constants.py
@@ -1,0 +1,19 @@
+COMMAND_PREFIX = "!"
+ACTIVE_GUILD = "oSTEM Global"
+
+AFFINITY_GROUPS = {
+    'womxn': ['womxn', 'women', 'woman'],
+    'acearo': ['ace', 'aro', 'ace/aro', 'acearo'],
+    'aapi': ['aapi'],
+    'disability': ['disability', '(dis)ability'],
+    'middle-sexualities': ['middle sexualities', 'middle-sexualities'],
+    'race-ethnicity': ['race and ethnicity', 'race', 'ethnicity', 'race-ethnicity'],
+    'transnon-binary': ['trans and non-binary', 'trans and nonbinary', 'trans', 'non-binary', 'enby', 'transgender', 'transnon-binary'],
+    'inqueery': ['inqueery']
+}
+
+WORKING_GROUPS = {
+    'dis-ability': ['disability', '(dis)ability'],
+    'beyondthebinary': ['binary', 'beyond the binary', 'beyond-the-binary', 'beyond_the_binary'],
+    'black_queer': ['black and queer', 'black', 'queery', 'black n queer', 'black_and_queer', 'black-and-queer', 'black_queer']
+}
